Main frontend completed . Admin in progress . Fixsing some bugs

+++ routes/web.php
@@ -2,45 +2,36 @@
 
 use Illuminate\Support\Facades\Route;
 
-/*
-|--------------------------------------------------------------------------
-| Web Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register web routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| contains the "web" middleware group. Now create something great!
-|
-*/
 
-Route::get('/', function (\App\Services\ProductService $productService) {
-    $products  = $productService->all();
-    return view('welcome',compact('products'));
-});
+Route::get('/', 'HomepageController');
 
 Auth::routes();
 Route::get('/home', 'HomeController@index')->name('home');
 Route::prefix('cart')->name('cart.')->group(function() {
+    Route::view('index','cart.index')->name('index');
     Route::post('add','CartController@add')->name('add');
     Route::delete('destroy/{id}','CartController@destroy')->name('destroy');
     Route::patch('update-quantity/{id}','CartController@updateQuantity')->name('update.quantity');
 });
 Route::name('categories.')->prefix('categories')->group(function (){
-    Route::get('/','CategoryController@index')->name('index');
+    Route::get('/','CategoryController@index')->middleware(['auth','admin'])->name('index');
+    Route::get('create','CategoryController@create')->middleware(['auth','admin'])->name('create');
     Route::post('store','CategoryController@store')->middleware('auth')->name('store');
     Route::get('show/{id}','CategoryController@show')->name('show');
-    Route::put('update/{id}','CategoryController@update')->name('update');
-    Route::delete('destroy/{id}','CategoryController@destroy')->name('destroy');
+    Route::get('edit/{id}','CategoryController@edit')->middleware(['auth','admin'])->name('edit');
+    Route::put('update/{id}','CategoryController@update')->middleware(['auth','admin'])->name('update');
+    Route::delete('destroy/{id}','CategoryController@destroy')->middleware(['auth','admin'])->name('destroy');
 });
 Route::name('products.')->prefix('products')->group(function (){
    Route::get("/", "ProductController@index")->name("index");
-   Route::get('create', 'ProductController@create')->middleware('auth')->name('create');
+   Route::get('create', 'ProductController@create')->middleware(['auth','admin'])->name('create');
    Route::post('store', "ProductController@store")->middleware('auth')->name('store');
    Route::get('show/{id}','ProductController@show')->name('show');
-   Route::get("edit/{id}","ProductController@edit")->middleware('auth')->name('edit');
+   Route::get("edit/{id}","ProductController@edit")->middleware(['auth','admin'])->name('edit');
    Route::put('update/{id}','ProductController@update')->middleware('auth')->name('update');
    Route::delete('destroy/{id}','ProductController@destroy')->middleware('auth')->name('destroy');
    Route::get('show-by-category/{categoryId}','ProductsByCategory')->name("by-category");
+
 });
 Route::name("orders.")->prefix("orders")->middleware("auth")->group(function (){
     Route::name("index")->get("/",'OrderController@index');
@@ -48,4 +39,8 @@ Route::name("orders.")->prefix("orders")->middleware("auth")->group(function (){
     Route::post('store','OrderController@store')->name('store');
     Route::put("update/{id}",'OrderController@update')->name("update");
     Route::delete("destroy/{id}",'OrderController@destroy')->name("destroy");
+});
+Route::prefix('dashboard')->middleware(['auth','admin'])->name('dashboard.')->group(function(){
+    Route::get('main','DashboardStaticsController')->name('index');
+    Route::get('products','ProductDashboardIndex')->name('products');
 });
